style: move files from toolbelt dir and add BROWSER_PROVIDER_NAME constant (closes #70)
