+++ plugin/idg/browser/__init__.py
@@ -1,0 +1,5 @@
+from .browser import IdgProvider  # noqa: F401
+from .network_manager import NetworkRequestsManager  # noqa: F401
+from .tree_node_factory import DownloadAllIdgFilesAsync  # noqa: F401
+from .tree_node_factory import DownloadDefaultIdgListAsync  # noqa: F401
+from .remote_platforms import RemotePlatforms  # noqa: F401
