+++ plugin/idg/browser/network_manager.py
