+++ plugin/idg/toolbelt/__init__.py
@@ -2,8 +2,4 @@
 from .log_handler import PlgLogger  # noqa: F401
 from .preferences import PlgOptionsManager  # noqa: F401
 from .translator import PlgTranslator  # noqa: F401
-from .plugin_globals import PluginGlobals  # noqa: F401
 from .singleton import Singleton  # noqa: F401
-from .browser import IdgProvider  # noqa: F401
-from .remote_platforms import RemotePlatforms  # noqa: F401
-from .network_manager import NetworkRequestsManager  # noqa: F401
