+++ plugin/idg/toolbelt/preferences.py
@@ -153,4 +153,6 @@ class PlgOptionsManager:
 
         settings.endGroup()
 
-        idg.toolbelt.PluginGlobals.instance().reload_globals_from_qgis_settings()
+        from idg.plugin_globals import PluginGlobals
+
+        PluginGlobals.instance().reload_globals_from_qgis_settings()
